pyproject.toml
[project]
name = "cir"
version = "0.2.3"
description = "Collection Instrument Registry"
requires-python = ">=3.13,<3.14"
dependencies = [
    "coverage==7.11.0",
    "fastapi==0.121.0",
    "firebase-admin==7.3.0",
    "gcp-storage-emulator==2024.8.3",
    "google-cloud-pubsub==2.36.0",
    "httpx==0.28.1",
    "mock-firestore==0.11.0",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "pip-audit==2.10.0",
    "protobuf==6.33.5",
    "pydantic==2.12.3",
    "pydantic-settings==2.13.1",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "pyyaml==6.0.3",
    "ruff==0.14.3",
    "setuptools==80.9.0",
    "starlette==0.49.3",
    "uvicorn==0.38.0",
    "sds-common==1.0.17"
]

[[tool.uv.index]]
name = "sds-repo"
url = "https://europe-west2-python.pkg.dev/ons-sds-ci/sds-python-packages/simple/"

[dependency-groups]
version-check = [
 "toml>=0.10.2",
 "tomlkit>=0.13.3",
]


[tool.ruff]
line-length = 120
indent-width = 4
exclude = ["tests/integration_tests", "tests/unit_tests", "tests/test_data", "scripts"]

[tool.ruff.lint]
select = [
    # Enabling ALL is not recommended it will implicitly enable new rules after upgrade.
    # "ALL",

    # Ruff rules: https://docs.astral.sh/ruff/rules/

    "E",      # pycodestyle erros
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "I",      # isort
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "S",      # flake8-bandit
    #"D",      # pydocstyle - Enforce existing docstrings only
    "C90",    # mccabe
    "RUF",    # Ruff specific rules

    # PL - Pylint is only partially supported, we also use the pylint tool to catch all the rules.
    # It is enabled here to take advantage of the ruff's speed.
    "PL",
]

ignore = [
    # Conflicts with google docstring style
    "D205",
    # Allow missing docstring, remove to enforce docstrings across the board
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    # Disable line length check as it is handled by black
    # :TODO: Remove E501 when ruff supports all black rules
    "E501",
    # indentation contains tabs
    "W191",
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    # Whitelisted default arguments
    "fastapi.Depends",
    "fastapi.Query",
]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    # Allow use of assert statements in tests
    "S101",
    # Allow use of auth_token - false positive for hardcoded password
    "S105",
    # Default timeout is added to kwargs
    "S113",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
